Rename old controllers to v1
