+++ app/controllers/v1/application_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# Superclass for all V1 controllers
+module V1
+  class ApplicationController < ::ApplicationController
+  end
+end
