+++ app/controllers/v1/rule_results_controller.rb
@@ -2,6 +2,15 @@
 
 # API for RuleResults
 module V1
-  class RuleResultsController < ::RuleResultsController
+  class RuleResultsController < ApplicationController
+    def index
+      render json: RuleResultSerializer.new(scope_search, metadata)
+    end
+
+    private
+
+    def resource
+      RuleResult
+    end
   end
 end
