+++ app/controllers/v1/rules_controller.rb
@@ -2,6 +2,41 @@
 
 # REST API for Rules
 module V1
-  class RulesController < ::RulesController
+  class RulesController < ApplicationController
+    def index
+      render json: RuleSerializer.new(scope_search, metadata)
+    end
+
+    def show
+      rule = if ::UUID.validate(params[:id])
+               search_by_id
+             else
+               search_by_ref_id
+             end
+
+      raise ActiveRecord::RecordNotFound if rule.blank?
+
+      render json: RuleSerializer.new(rule)
+    end
+
+    private
+
+    def resource
+      Rule
+    end
+
+    def search_by_id
+      ::Pundit.policy_scope(User.current, ::Rule).friendly.find(params[:id])
+    end
+
+    def search_by_ref_id
+      rule = Rule.latest.where(
+        'rules.slug LIKE ?',
+        "%#{ActiveRecord::Base.sanitize_sql_like(params[:id])}%"
+      )
+      raise ActiveRecord::RecordNotFound if rule.blank?
+
+      rule.first
+    end
   end
 end
