+++ app/controllers/v1/systems_controller.rb
@@ -2,6 +2,22 @@
 
 # API for Systems (only Hosts for the moment)
 module V1
-  class SystemsController < ::SystemsController
+  class SystemsController < ApplicationController
+    def index
+      render json: HostSerializer.new(scope_search, metadata)
+    end
+
+    def destroy
+      host = Host.find(params[:id])
+      authorize host
+      host.destroy
+      render HostSerializer.new(host)
+    end
+
+    private
+
+    def resource
+      Host
+    end
   end
 end
