+++ test/controllers/v1/profiles_controller_test.rb
@@ -5,7 +5,7 @@ require 'test_helper'
 module V1
   class ProfilesControllerTest < ActionDispatch::IntegrationTest
     setup do
-      ::ProfilesController.any_instance.stubs(:authenticate_user)
+      ProfilesController.any_instance.stubs(:authenticate_user)
       User.current = users(:test)
       users(:test).update! account: accounts(:test)
       profiles(:one).update! account: accounts(:test)
