+++ test/controllers/v1/systems_controller_test.rb
@@ -5,11 +5,11 @@ require 'test_helper'
 module V1
   class SystemsControllerTest < ActionDispatch::IntegrationTest
     setup do
-      ::SystemsController.any_instance.expects(:authenticate_user)
+      SystemsController.any_instance.expects(:authenticate_user)
     end
 
     test 'index lists all systems' do
-      ::SystemsController.any_instance.expects(:policy_scope).with(Host)
+      SystemsController.any_instance.expects(:policy_scope).with(Host)
                          .returns(Host.all).at_least_once
       get v1_systems_url
 
@@ -17,7 +17,7 @@ module V1
     end
 
     test 'index accepts search' do
-      ::SystemsController.any_instance.expects(:policy_scope).with(Host)
+      SystemsController.any_instance.expects(:policy_scope).with(Host)
                          .returns(Host.all).at_least_once
       get v1_systems_url, params: { search: 'name=bar' }
